feat: implement user management functionality with CRUD operations and role-based access control

+++ backend/src/controllers/auth-controller/logout-controller.ts
@@ -2,25 +2,23 @@ import { NextFunction, Request, Response } from "express";
 import { tokenService } from "../../mongo/auth/token-service";
 
 export async function logoutController(
-  req: Request,
+  req: Request & { user?: { id: string } },
   res: Response,
   next: NextFunction
 ) {
   try {
-    const loggedinUser = req.user;
-    console.log("logged in user", loggedinUser);
-
+    const token = req.cookies.authorization;
+    const userId = req.user?.id || "";
     res.clearCookie("authorization");
+    if (token && userId) {
+      await tokenService.deleteToken({ userId, token });
+    }
 
-    await tokenService.deleteToken({
-      userId: loggedinUser?.id || "",
-      token: req.cookies.authorization,
-    });
     res.status(200).json({
       message: "You are logged out",
     });
   } catch (error) {
-    console.error(error);
+    console.error("Logout error: ", error);
     next({
       status: 500,
       message: (error as Error).message,
